pyproject.toml
[tool.poetry]
name = "hubspot-utils"
version = "0.1.0"
description = ""
authors = ["Haris Ibrahim K. V. <blucalvin@gmail.com>"]
readme = "README.md"
packages = [{include = "hubspot_utils"}]

[tool.poetry.dependencies]
python = "^3.11"
hubspot-api-client = "^9.0.0"
aws-lambda-powertools = "^3.0.0"
botocore = "^1.35.31"
boto3 = "^1.35.31"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
